+++ force-app/main/default/lwc/validationCard/validationCard.html
@@ -10,12 +10,16 @@
         onclick={deleteValidation}
       />
     </span>
+
     <span class="gray">{label.question}: </span
     ><span>{validation.Related_Question__c.Label__c}</span><br />
+
     <span class="gray">{label.dependent_question}: </span
     ><span>{validation.Dependent_Question__c.Label__c}</span><br />
+
     <span class="gray">{label.operator}: </span
     ><span>{validation.Operator__c}</span><br />
+
     <span class="gray">{label.compared_value}: </span
     ><span>{validation.Value__c}</span><br />
   </div>
