+++ force-app/main/default/lwc/validationForm/validationForm.html
@@ -1,0 +1,79 @@
+<template>
+  <div class="slds-box slds-box_x-small slds-m-around_x-small">
+    <div class="slds-grid slds-wrap">
+      <div class="slds-size_1-of-5">
+        <div class="slds-m-around_x-small">
+          <lightning-combobox
+            label={label.question}
+            value={firstPosition}
+            options={questionOptions}
+            onchange={selectFirstQuestion}
+            class="firstCombobox"
+          >
+          </lightning-combobox>
+        </div>
+      </div>
+
+      <div class="slds-size_1-of-5">
+        <div class="slds-m-around_x-small">
+          <lightning-combobox
+            label={label.dependent_question}
+            value={secondPosition}
+            options={questionOptions}
+            onchange={selectSecondQuestion}
+            class="secondCombobox"
+          >
+          </lightning-combobox>
+        </div>
+      </div>
+
+      <div class="slds-size_1-of-5">
+        <div class="slds-m-around_x-small">
+          <lightning-combobox
+            label={label.operator}
+            value={selectedOperator}
+            placeholder={label.select_operator}
+            options={displayedOperators}
+            class="slds-m-around_x-small"
+            onchange={setSelectedOperator}
+          >
+          </lightning-combobox>
+        </div>
+      </div>
+
+      <div class="slds-size_1-of-5">
+        <div class="slds-m-around_x-small">
+          <lightning-combobox
+            label={label.value}
+            if:true={isFirstQuestionPicklist}
+            placeholder={label.select_value}
+            options={questionsOptions}
+            class="input"
+            required
+          >
+          </lightning-combobox>
+
+          <lightning-input
+            if:false={isFirstQuestionPicklist}
+            type={inputType}
+            label={label.enter_compared_value}
+            class="input"
+            required
+          ></lightning-input>
+        </div>
+      </div>
+
+      <div class="slds-size_1-of-5">
+        <div class="slds-m-top_x-large">
+          <div class="slds-m-left_x-small">
+            <lightning-button
+              variant="brand"
+              label={label.add}
+              onclick={addValidation}
+            ></lightning-button>
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
+</template>
